requirements.txt
pandas~=2.2
numpy~=1.26
openpyxl
sqlalchemy~=2.0
psycopg2-binary~=2.9
requests~=2.32
beautifulsoup4~=4.13.3
google-auth~=2.36
google-api-python-client~=2.152
pytest~=8.
pytest-cov~=4.1
